Get wall posts from Facebook
Pull /<id>/feed instead of /<id>/posts so wall posts by other users are
retrieved along with posts by the user.

Wall posts set the in_reply_to_user_id field on the post.

Data files for /feed API calls were copied from /posts files, with a
wall post added to Gina's feed.

The dashboard for Facebook is called Wall Posts.  For Twitter, it's
called "Messages to you" for lack of a better name, and is also messages
sent to a user, but not in response to a previous message.